[HOMEWORK-1] UI refresh on edit button click fixed, unnecessary useCallback removed, CreateNote and EditNode components merged
